requirements.txt
# Core dependencies
Pillow>=9.0.0
requests>=2.28.0
python-dotenv>=0.19.0
schedule>=1.1.0
psutil>=5.8.0

# Web interface
flask>=2.0.0
flask-cors>=3.0.10

# Voice control (optional)
pyttsx3>=2.90
SpeechRecognition>=3.8.1
pyaudio>=0.2.11

# AI integration (optional)
openai>=0.27.0

# Raspberry Pi hardware (optional - only needed on Pi)
RPi.GPIO>=0.7.1
spidev>=3.5

# E-ink display library (optional - install manually if needed)
# IT8951>=0.1.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0